--- src/components/atoms/icons/index.ts
@@ -1,4 +1,0 @@
-export * from './CheckIcon'
-export * from './Icon'
-
-export type { CheckIconProps } from './types'
