--- src/components/atoms/icons/types.ts
@@ -1,5 +1,0 @@
-export type CheckIconProps = {
-  width?: number
-  height?: number
-  color?: string
-}
