+++ src/components/atoms/index.ts
@@ -3,7 +3,7 @@ export { LinearGradient as Gradient } from 'expo-linear-gradient'
 
 export * from './Absolute'
 export * from './AbsoluteFullFill'
-export * from './icons'
 export * from './Input'
 export * from './Spacer'
 export * from './types'
+export * from './Icon'
