+++ src/navigation/BottomTabNavigator.tsx
@@ -1,11 +1,12 @@
-import { Feather } from '@expo/vector-icons'
 import { createBottomTabNavigator, BottomTabNavigationOptions } from '@react-navigation/bottom-tabs'
 import { FC } from 'react'
 
 import { ExamplesStack } from './ExamplesStack'
 import { HomeStack } from './HomeStack'
 
+import { Icon } from '~components'
 import { useCallback, useNavigationTheme, useTranslation } from '~hooks'
+import { IconNames } from '~types/icon'
 
 const { Navigator, Screen } = createBottomTabNavigator<MainTabParamList>()
 
@@ -17,19 +18,19 @@ export const BottomTabNavigator: FC = () => {
 
   const screenOptions = useCallback<ScreenOptions>(
     ({ route }) => ({
-      tabBarIcon: ({ color, size }) => {
-        let iconName: keyof typeof Feather.glyphMap
+      tabBarIcon: ({ color, size, focused }) => {
+        let iconName: IconNames
 
         if (route.name === 'HomeStack') {
-          iconName = 'home'
+          iconName = focused ? 'home-5-fill' : 'home-line'
         } else if (route.name === 'ExamplesStack') {
-          iconName = 'list'
+          iconName = focused ? 'file-list-2-fill' : 'file-list-2-line'
         } else {
-          iconName = 'alert-triangle'
+          iconName = 'alert-line'
         }
 
         // You can return any component that you like here!
-        return <Feather name={iconName} size={size} color={color} />
+        return <Icon name={iconName} size={size} color={color} />
       },
       headerShown: false,
       ...tabBarTheme,
