+++ src/screens/ExamplesScreen.tsx
@@ -1,6 +1,5 @@
 import { Button, ScrollView, Center } from 'native-base'
 
-import { Icon } from '~components'
 import { useTranslation, useCallback } from '~hooks'
 
 export const ExamplesScreen = (props: ExamplesScreenProps): JSX.Element => {
@@ -48,7 +47,6 @@ export const ExamplesScreen = (props: ExamplesScreenProps): JSX.Element => {
         <Button size="lg" width="64" mb={2} onPress={goToCityListScreen_EXAMPLE}>
           {t('examples_screen.go_to_screen_with_BEdata')}
         </Button>
-        <Icon name="account-box-fill" size={24} color={'amber.600'} />
       </Center>
     </ScrollView>
   )
